+++ Character_header/Dragon.h
@@ -11,7 +11,7 @@ private:
     // Powers specialpowers;
 
 public:
-    Dragon(const string &name, double health, double attackStrength, const string &allegiance);
+    Dragon(const string &name, int health, int attackStrength, const string &allegiance);
     void fly(string);
     int getfireAttack();
     int getfireBall();
